pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slideforge"
version = "0.1.0"
description = "Python slide generation engine for academic and AI-assisted presentations"
requires-python = ">=3.10"
dependencies = [
    "python-pptx",
    "matplotlib",
    "numpy",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]